refactors & reorg. for modularity
